+++ CHANGELOG.md
@@ -12,6 +12,12 @@ The API is unstable, and [Semantic Versioning](https://semver.org/spec/v2.0.0.ht
 		Click to see more.
 	</summary>
 
+Nothing here yet.
+
+</details>
+
+## [0.4.1] - 2021-10-20
+
 ### Fixed
 - Ability to use `MenuBar` without `$Window.js` (`TypeError: Assignment to constant variable.`)
 - Handle `document.body` not existing if you create a `MenuBar` before `DOMContentLoaded`
@@ -153,7 +159,8 @@ The API is unstable, and [Semantic Versioning](https://semver.org/spec/v2.0.0.ht
 - Window component is now an app window instead of a tool window, to aid reintegration with 98.js.org; jspaint will come later.
 - Window component is styled with `.os-window` now, altho it includes both classes `os-window` and `window`.
 
-[Unreleased]: https://github.com/1j01/os-gui/compare/v0.4.0...HEAD
+[Unreleased]: https://github.com/1j01/os-gui/compare/v0.4.1...HEAD
+[0.4.1]: https://github.com/1j01/os-gui/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/1j01/os-gui/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/1j01/os-gui/compare/v0.2.2...v0.3.0
 [0.2.2]: https://github.com/1j01/os-gui/compare/v0.2.1...v0.2.2